Preserve template git history (#6)
The git history of the template repository was being lost because the
`gh repo create --template` option would squash the history to a single
commit.

The script has been updated to avoid using the `--template` option.
Instead the template repository is checked out locally. This lets us
push the entire git history to each new repository.

The script will also ask for any PR head branches, and push each of
those up to the new repo as well.