Preserve issue order (#9)
The setup script now ensures that issues and PRs are created in the
same order they are in on the template repository. It has also been
updated to create issues first, rather than PRs, because we're now
using issue number references in various places (and we replaced the
one PR reference we had beforehand).