Convert the setup script to a GitHub action
This PR converts this repository from a "setup repo" for the shared
libraries technical challenge to being a reusable GitHub action for any
team to create per-candidate technical challenge repositories. We will
clone the repo and split the git history in two here.

The README should explain the purpose of the action and how to set it
up and use it.